Breakdown Events

+++ Assets/Supplies.cs
@@ -7,7 +7,7 @@ public class Supplies : MonoBehaviour
     //use one food per person per move
     public int Food = 80;
     //I'm assuming you have to do 10 stops and each one takes 3-5 fuel depening on distance (in game default is 5)
-    public int Fuel = 60;
+    public float Fuel = 60;
     //Hull Parts
     public int RepairHull = 1;
     //Electronics
